MyPro/requirement.txt
asgiref==3.4.1
Django==3.2.15
django-crispy-forms==1.13.0
django-environ==0.9.0
djangorestframework==3.13.1
pkg-resources==0.0.0
pytz==2022.2.1
sqlparse==0.4.2
typing-extensions==4.1.1
